This is the video of the working physical breadboard.
The first time running through the code I demonstrate how the buttons cannot be pressed outside of their specific order. And the second time through I go all the way through it without pressing the buttons out of order. Also added was a way to reset the buttons without resetting the entire board. This was utilized to reset the board for the second test.